OGV and SRT are generic formats, not Dark Forces formats, so move them to TFE_Asset
